a1

+++ pybricks/pyBricks1/main.py
@@ -66,7 +66,7 @@ async def get_elements_at_greenhouse():
     await GripperController.grip_element_using_both_arms()
     await WheelController.move_wheels_forward_in_straight_line(float(30))
     await WheelController.wheel_u_turn_right()
-    await WheelController.move_wheels_forward_in_straight_line(float(50))
+    await WheelController.move_wheels_forward_in_straight_line(float(40))
     await GripperController.reset_both_arms()
     await WheelController.wheel_right_turn_with_angle(float(20))
     await WheelController.move_wheels_forward_in_straight_line(float(55))
@@ -120,7 +120,7 @@ async def get_the_vegetables_at_the_market():
     await WheelController.wheel_left_turn()
 
     # go to the 2 new vegetables
-    await WheelController.move_wheels_forward_in_straight_line(float(290))
+    await WheelController.move_wheels_forward_in_straight_line(float(260))
     await WheelController.wheel_left_turn()
     await WheelController.move_wheels_forward_in_straight_line(float(50))
     await WheelController.wheel_left_turn_with_angle(20)
